Add clear cache admin menu item
