+++ home/templates/wagtailadmin/clear_cache.html
@@ -1,0 +1,12 @@
+{% extends "wagtailadmin/base.html" %}
+
+{% block titletag %}Commands: Clear Cache{% endblock %}
+
+{% block content %}
+    {% include "wagtailadmin/shared/header.html" with title="Clear Cache" icon="placeholder" %}
+    <section class="nice-padding">
+        <form action="" method="post">{% csrf_token %}
+            <input type="submit" value="Clear Cache Now" class="button" />
+        </form>
+    </section>
+{% endblock %}
